Added transformations. There are 4 new files: misc/kigtransform.* and objects/transform.*; the first collects all 'transformation-related' stuff, the other collects all 'object-to-transform-related' stuff.
Right now the following objects can be transformed:
point, line, ray, segment, circle, conic, cubic.

There is support for the following transformations:
- central symmetry about a point,
- reflection about a line,
- rotation of a given angle about a point,
- translation of a given vector,
- scaling of a given factor about a given point,
- scaling of a given factor about a given line (this is the first example
of a non-homothetic transformation, i.e. a circle does not map onto a
circle),
- an example of a projective transformation that is not an affine map:
it depends on a ray and an angle.

The addition of the transformations is extremely localized in those files:
changes are present only in kig_part.cpp in order to add the new types
(one per object that can be transformed). Right now there is no new entry
in kigpartui.rc, anyway the transformations can be reached by means of the
contextual menu (right click) of any object.

CCMAIL: devriese@kde.org

svn path=/trunk/kdeedu/kig/; revision=195210