Introducing the "CastShadow" transformation.  This is a projective transformation that computes the "shadow" of a given object from a given light source onto a plane.
The plane is given by a line that indicates the intersection with the
"vertical" plane containing all the graphic objects.

The light source should be intended as being "nearer" to the eye with
respect to the objects plane.

Right now there is an extra degree of freedom that is arbitrarily
decided by the transformation, which actually is the distance of the light
source from the object plane.  Perhaps this can be changed in the future.

This transformation can be simulated by hands as follows:

To transform a point P using a light source A and a line l:

1. Construct the "mirror" image of A with respect to l, call it B.

2. Project P orthogonally onto the line l, call this point H.

3. draw the line through A and P, call it r.

4. draw the line through B and H, call it s.

5. intersect r and s, the intersection point Q is the result of the
   transformation.

CCMAIL: devriese@kde.org

svn path=/branches/make_kdeedu_cool/kdeedu/kig/; revision=205988