1. Patch suggested by domi in order to tag an invalid coordinate.  This would imply many simplifications (e.g. we could drop the "valid" bool in getPoint), which could be done in the future; by now only drawLocus takes advantage of the change.
2. Small changes in getPoint for cubics, with less strict "asserts"
because we must accept situations where the newton iterations do not
converge; moreover now not all parameters in [0,1] lead to valid
points on the cubic.  In the previous version all parameters mapped
to valid coordinates, but the mapping was not injective (different
parameters giving the same point).

(mp)
CCMAIL:devriese@kde.org

svn path=/trunk/kdeedu/kig/; revision=222293