As discussed with Pino, polygons are drawn by filling them rather then drawing their boundary.
At the moment they can be selected by clicking on the boundary;
it is not obvious that selection should be done by clicking in
the inside, since then one would perhaps lose the possibility
of selecting objects inside the polygon unless a proper priority
is enforced.  So by now I leave the selection as is.
This is however inconvenient when dealing with the selection of
segments composing the boundary.

svn path=/trunk/kdeedu/kig/; revision=373680