Lots of math handling.
<cmath> doesn't necessarily give us all the math.h names
in the global namespace; we can fix that by doing either

include <math.h>
using std::
std::

to the code. It varies a little. With constants like M_PI,
we need <math.h>. Otherwise, I've tried to use the ones that
cause the least change.

Marc Mutz previously pointed out that we might have compilers that
do *not* put the math things in namespace std::, in which case
we're screwed anyway. I'm not aware of any such, though (will code
for awareness-expanding drugs, for sure).

svn path=/trunk/KDE/kdeedu/kig/; revision=791947