Avoid processing parameters if the number of arguments is not met
Changed the finished variable to finish, as I think is the way the code
was intended to be. Also fixes the second validation that took place
(calling canFinish in the finishPressed method) since first validation
happens in the validatePage method of the Arguments page.
FIXED-IN: 4.7.3
CCBUG: 282913
REVIEW: 102725

cherry-picked from 7301b06a30b994210e713950c8600e0f25558a80