Implemented the Parsing Mechanism for Geogebra
Based on XSLT 2.0 and QtXmlPatterns. The XSLT provided is just a proof
of concept that should be fleshed out - and the code should be modified
accordingly to handle all the different types of objects.

I am actually not sure every possible object can be built using this
mechanism but many objects can be created by just selecting the object
type and stacking the parent objects so this should get us far.