Conrete Example of the XSLT Implementation
This is a better example of how the XLST will look like. The idea is to
add templates for the different object types and call these from the
xsl:for-each statement at the main template depending on the type and
attributes of the nodes.

So far supporting only FixedPoint and SegmentAB constructions.