Don't Create Duplicated Objects
The format of the Geogebra XML can lead to the creation of more than
one object with the same label because they are defined more than once.
We abort the creation of the object at the last minute, which may be less
efficient but is less intrusive in the code.