Fixed Startup as Unique Application
Added a dependency to DBusAddons framework to do the unique application
dance, which required a rework on KigApplication, making it a full
Q_OBJECT for the newInstance slot.

The startup needs to be fixed, probably by porting to QCommanLineArgs,
because the subsequent instances are not handling their command line
arguments properly.