Beef Up the Default Dir Logic for Opening Files
The port introduced a regression where the Open File Dialog no longer
started on the Documents folder by default. The new behavior restores
this by default, but also takes you to the previously opened file, if
any, which is helpful if you are opening several files from the same
folder.