Start a New Kig Application for the New Action
The previous behavior made sense when Kig was a KUniqueApplication, but
now that it is not, running a new application is the best way to go.
Another option would be to turn the Kig interface into an MDI, but that
requires a lot more work. Notice that the binary found to r.un Kig might
not be the same as the running binary.

REVIEW: 122497