Removing coordinatevalidator Test Case
There is no point in having a test case that is a copy of the original
code because it will not be updated when we change the original code (as
I recently did). I am keeping the source file around for reference as
testing should be the next big thing in Kig after the port is finished.