Optional Migration Code for KF5
Summary:
This is long overdue and the fact that it has not caused more trouble in
bugzilla means it is probably a non-issue, but just to be complete,
here's code that migrates the old configuration to the new directories.

BUG: 353729

Test Plan:
Delete the kigrc config file, and the kig/kig-types/macros.kigt files if you have them.
Run the application.
Run it with existing kig/kig-types/macros.kigt too to make sure it does not delete the files.

Reviewers: #kde_edu, apol

Reviewed By: apol

Subscribers: cfeck, apol

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D8225